requirements.txt
pip==23.2.1
wheel==0.41.2
setuptools==68.2.0
certifi==2023.11.17
idna==3.6
urllib3==2.1.0
requests==2.31.0
slack_sdk==3.26.1
load_dotenv==0.1.0
Flask==3.0.0
python-dotenv==1.0.0
jira==3.6.0